Chaged some layout and moved progressmenu
