+++ frontend/layouts/OnboardingLayout/ProgressBarMenu/ProgressBarMenu.jsx
