+++ frontend/layouts/OnboardingLayout/ProgressBarMenu/index.js
