+++ frontend/layouts/OnboardingLayout/ProgressBarMenu/styles.js
